Add precision in function help that `output_file` is expected to be a filename
`quarto render --output` will throw the error if the value provided is not supported. We don't want to add its own validation system in the package. If this is not clear enough, we could try to catch the error and rethrow differently, but this would still mean we need to maintain two error system for same usage.

Closes #89